+++ ssl/s3_pkt.c
@@ -469,7 +469,7 @@ int ssl3_write_bytes(SSL *s, int type, const void *buf_, int len)
      * accomodate up to 8 records, but the compromise is considered worthy.
      */
     if (type == SSL3_RT_APPLICATION_DATA &&
-        len >= 4 * (max_send_fragment = s->max_send_fragment) &&
+        len >= 4 * (int)(max_send_fragment = s->max_send_fragment) &&
         s->compress == NULL && s->msg_callback == NULL &&
         SSL_USE_EXPLICIT_IV(s) &&
         EVP_CIPHER_flags(s->enc_write_ctx->cipher) &
@@ -493,7 +493,7 @@ int ssl3_write_bytes(SSL *s, int type, const void *buf_, int len)
             if (packlen <= 0)
                 return -1;
 
-            if (len >= 8 * max_send_fragment)
+            if (len >= 8 * (int)max_send_fragment)
                 packlen *= 8;
             else
                 packlen *= 4;
@@ -545,7 +545,7 @@ int ssl3_write_bytes(SSL *s, int type, const void *buf_, int len)
                                           EVP_CTRL_TLS1_1_MULTIBLOCK_AAD,
                                           sizeof(mb_param), &mb_param);
             
-            if (packlen <= 0 || packlen > wb->len) { /* never happens */
+            if (packlen <= 0 || packlen > (int)wb->len) { /* never happens */
                 free(wb->buf); /* free jumbo buffer */
                 wb->buf = NULL;
                 break;
